+++ karma.conf.js
@@ -1,8 +1,6 @@
 // Karma configuration
 // Generated on Wed Mar 08 2017 13:05:58 GMT-0800 (PST)
 
-var isTravis = !!process.env.TRAVIS;
-
 module.exports = function(config) {
   var configuration = {
 
@@ -77,13 +75,17 @@ module.exports = function(config) {
     coverageReporter: {
       type : 'text',
       dir : 'coverage/'
+    },
+
+    customLaunchers: {
+      'FirefoxHeadless': {
+        base: 'Firefox',
+        flags: [
+          '-headless',
+        ],
+      }
     }
   };
-
-  // run code coverage locally
-  if (!isTravis) {
-    configuration.reporters.push('coverage');
-  }
 
   config.set(configuration);
 };